Updating to use existing send-block-kit-message key
